Compute the origin and scale of the horizon plot

Compute the origin and scale of the horizon plot

compute_ori_sca(data, origin, horizonscale)

Arguments

data

A data frame.

origin

Origin of the horizon plot. It can either be a string, namely "midpoint" (the default), "median", "mean", "min" or "quantiles", or a user-specified number.

horizonscale

Cutpoints of the horizon plot. It can either be an integer specifying the number of ranges (default is 6), or a user-supplied numeric vector with the cutpoints defining the different ranges.

Value

A list with two elements: ori, which represents the numeric origin of the horizon plot, and vec_cut, which is a vector with the cutpoints for the horizon plot.